requirements.txt
requests==2.31.0
pyjwt==2.6.0
vantage6-client==3.10.4
pandas==1.5.3
scipy==1.11.1
scikit-learn==1.3.0
flake8==6.1.0